+++ sdk/dotnet/GetDatabase.cs
@@ -83,6 +83,10 @@ namespace KomminarLabs.InfluxDB3
         /// </summary>
         public readonly string Name;
         /// <summary>
+        /// The template partitioning of the cluster database.
+        /// </summary>
+        public readonly ImmutableArray<Outputs.GetDatabasePartitionTemplateResult> PartitionTemplates;
+        /// <summary>
         /// The retention period of the cluster database in nanoseconds.
         /// </summary>
         public readonly int RetentionPeriod;
@@ -101,6 +105,8 @@ namespace KomminarLabs.InfluxDB3
 
             string name,
 
+            ImmutableArray<Outputs.GetDatabasePartitionTemplateResult> partitionTemplates,
+
             int retentionPeriod)
         {
             AccountId = accountId;
@@ -109,6 +115,7 @@ namespace KomminarLabs.InfluxDB3
             MaxColumnsPerTable = maxColumnsPerTable;
             MaxTables = maxTables;
             Name = name;
+            PartitionTemplates = partitionTemplates;
             RetentionPeriod = retentionPeriod;
         }
     }